Skip failing metacharacter scenario in Travis (refs #7676)
We have had a long-standing failing test in Travis regarding metacharacter
filtering. However, this test passes oddly in various environments - most
notably, in the End Point devcamps environment, which we're no longer using.
As far as I can tell, the feature is working as expected in our production
environment, and isolating the test's issues has proven difficult.

Rather than removing the test entirely, this commit adds a Cucumber 'tag'
to the scenario, and changes the default set of tests run by Cucumber
in Travis. If an environment variable named 'TRAVIS' is present and equals
'true' (as a string, not as a boolean), Cucumber will automatically skip
any tests tagged with '@travis-exclude'.

See #7676 and #5628.