Add parameter for exact matches of fields
Fixes https://issues.dp.la/issues/8398 ("Faceting is inaccurate").

Adds a querystring parameter to the API named `exact_field_match`, which
disables tokenized searching on fields given in the `fields' parameter.
These fields are used in applications like the DPLA frontend where the
user follows links based on facet values from previous searches.  A
tokenized search -- for example, where the phrase "University of
Pennsylvania" is broken down and the terms "University" and
"Pennsylvania" are considered indepenedently -- is not wanted in this
case, where we need an exact filter placed on "University of
Pennsylvania".

Also add some documentation comments in various places, which came out
of analyzing the code prior to making my changes.